Fix clippy lint errors in study modules
- cost_model.rs: Extract MeasuredEffects struct to reduce for_scale()
  from 8 args to 2 (clippy::too_many_arguments)
- noise.rs: Remove redundant as f64 cast on thermal_sigma_c
- stats.rs: Replace n % 2 == 0 with n.is_multiple_of(2)

All three crates now pass cargo clippy --all-targets -- -D warnings.

+++ crates/gpu-harness/src/study/noise.rs
@@ -79,7 +79,7 @@ impl NoiseModel {
 
     /// Add Gaussian noise to a temperature reading.
     pub fn jitter_thermal(&self, temp_c: f32, rng: &mut impl Rng) -> f32 {
-        let noise = self.gaussian(rng, 0.0, self.thermal_sigma_c as f64);
+        let noise = self.gaussian(rng, 0.0, self.thermal_sigma_c);
         (temp_c + noise as f32).max(0.0)
     }
